Fix structure for types
